+++ CHANGELOG.md
@@ -9,9 +9,9 @@
 
 ## Unreleased
 
+- Add plek as a dependency (PR #834)
 - Provide `GovukPublishingComponents.render` method for rendering components outside of views (PR #832)
 - Remove govspeak dependency with kramdown for markdown rendering (PR #827)
-- Add gds-adpi-adapters and plek as dependencies (PR #830)
 
 ## 16.13.0