+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
-## Unreleased
+## 16.14.0
 
 - Add attachment link (experimental) component (PR #833)
 - Add plek as a dependency (PR #834)